requirements.txt
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.11.0
drf-yasg==1.17.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
openapi-codec==1.3.2
packaging==20.0
pandas==0.24.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
whitenoise==4.1.3
